Prune deposit.md of 404 exchanges
OpenLedger is now OpenLedger and does not currently provide a fiat bridge (and being based in Denmark they are probably not going to in the near future either).

Many exhanges and bridges have delisted BTS or ceased to exist. It might be better to link to [Coinmarketcap markets section](https://coinmarketcap.com/currencies/bitshares/#markets) for the newer exchanges, but I've added Binance.